Backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
Werkzeug==2.3.7
scikit-learn==1.3.0
pandas==2.0.3
matplotlib==3.7.2
gunicorn==21.2.0 